Add sample CSVs from real output
Output is from running the model with the full dataset, for 30 rounds, and 10 clients sampled. They are meant to be used with create_charts_group.m.
Each run took about 30 minutes (Ryzen 7 7700X w/ 32GB of RAM, on Linux Mint 22.2 Zara through a venv set up in PyCharm), and produced two CSVs (which were renamed in accordance to the MATLAB script, with the "results" CSVs being sorted "in post").